requirements.txt
discord.py>=2.3.2
yt-dlp>=2024.3.10
requests>=2.31.0
python-dotenv>=1.0.1
